finished gui

+++ common/src/main/java/immersive_melodies/network/c2s/MelodyDeleteRequest.java
@@ -1,9 +1,13 @@
 package immersive_melodies.network.c2s;
 
 import immersive_melodies.cobalt.network.Message;
+import immersive_melodies.cobalt.network.NetworkHandler;
+import immersive_melodies.network.s2c.MelodyListMessage;
 import immersive_melodies.resources.ServerMelodyManager;
+import immersive_melodies.util.Utils;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.network.PacketByteBuf;
+import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.util.Identifier;
 
 public class MelodyDeleteRequest extends Message {
@@ -24,7 +28,10 @@ public class MelodyDeleteRequest extends Message {
 
     @Override
     public void receive(PlayerEntity e) {
-        // todo security
-        ServerMelodyManager.deleteMelody(identifier);
+        if (Utils.canDelete(identifier, e)) {
+            ServerMelodyManager.deleteMelody(identifier);
+
+            NetworkHandler.sendToPlayer(new MelodyListMessage(), (ServerPlayerEntity) e);
+        }
     }
 }
